[API-359] Implement non-blocking user score update function (#13245)
### Description
This is not the most ideal fix, but it's less dangerous than an overhaul
to the functions and tables involved.
A better long term fix would be to update get_user_scores() to require a
list of IDs, fetch those IDs in batches, and run the updates in batches
so that no part of this query takes several hundred seconds.

But this at least makes the very long running score computation
non-blocking for other writers by writing it into a temp table. In
practice, the number of user scores that get updated per round is a few
thousand at most.

### How Has This Been Tested?
Tested the function against live data and verified it runs without
issue. Will also make sure discovery integration tests are passing since
they check the update results.