[PE-7293] Prefer Meteora for external wallet swaps (#13374)
### Description

Updates external wallet swaps to use Meteora due to issues with Jup. The
approach is:
artist-coin <-> artist-coin uses meteora for quotes and swaps in a
single transaction
usdc/sol <-> artist-coin uses jupiter to handle usdc <-> audio, and
meteora for audio <-> artist-coin for quotes and swaps, which are in two
transactions.

### How Has This Been Tested?

Tested all forms of buy/sell/swap